requirements.txt
django>=1.8,<1.9
djangorestframework
PyMySQL==0.6.2
requests
MySQL-python